Try except en usuario

+++ api/usuario.py
@@ -16,26 +16,41 @@ def usuario():
 
 @ruta_usuario.route("/saveusuario", methods=["POST"])
 def saveusuario():
-    id_usuario = request.json['id_usuario']
-    nombre=request.json['nombre']
-    new_usuario= usuario(id_usuario,nombre)
-    db.session.add(new_usuario)
-    db.session.commit()
-    return "Datos guardados con exitos"
+    try:
+        id_usuario = request.json['id_usuario']
+        nombre=request.json['nombre']
+        new_usuario= usuario(id_usuario,nombre)
+        db.session.add(new_usuario)
+        db.session.commit()
+        return "Datos guardados con exitos"
+    except Exception as e:
+        return f"Hubo un error {str(e)}"
+    
 
 @ruta_usuario.route("/updateusuario", methods=["PUT"])
 def updateusuario():
-    id_usuario = request.json['id_usuario']
-    nombre=request.json['nombre']
-    nusuario = usuario.query.get(id_usuario) #Select * from Cliente where id = id
-    nusuario.nombre=nombre 
-    db.session.commit()
-    return "Datos Actualizado con exitos"
+    try:
+        id_usuario = request.json['id_usuario']
+        if not id_usuario:
+            return "El usuario no esta registrado"
+        nombre=request.json['nombre']
+        nusuario = usuario.query.get(id_usuario) #Select * from Cliente where id = id
+        nusuario.nombre=nombre 
+        db.session.commit()
+        return "Datos Actualizado con exitos"
+    except Exception as e:
+        return f"Hubo un error {str(e)}"
 
 @ruta_usuario.route("/deleteusuario/<id>", methods=["DELETE"])
 def deleteusuario(id):
-    id_usuario = request.json['id_usuario']
-    usuariox = usuario.query.get(id_usuario)
-    db.session.delete(usuariox)
-    db.session.commit()
+    try:
+        id_usuario = request.json['id_usuario']
+        if not id_usuario:
+            return "El usuario no se encuentra en la base de datos"
+        usuariox = usuario.query.get(id_usuario)
+        db.session.delete(usuariox)
+        db.session.commit()
+        return "El usuario ha sido eliminado con exito"
+    except Exception as e:
+        return f"Hubo un error{str(e)}"
